+++ ta4j-core/src/main/java/org/ta4j/core/analysis/Returns.java
@@ -92,9 +92,9 @@ public class Returns implements Indicator<Num> {
         this.type = type;
         // at index 0, there is no return
         values = new ArrayList<>(Collections.singletonList(NaN.NaN));
-        calculate(position);
+        calculate(position, barSeries.getEndIndex());
 
-        fillToTheEnd();
+        fillToTheEnd(barSeries.getEndIndex());
     }
 
     /**
@@ -111,7 +111,7 @@ public class Returns implements Indicator<Num> {
         values = new ArrayList<>(Collections.singletonList(NaN.NaN));
         calculate(tradingRecord);
 
-        fillToTheEnd();
+        fillToTheEnd(tradingRecord.getEndIndex(barSeries));
     }
 
     public List<Num> getValues() {
@@ -142,10 +142,6 @@ public class Returns implements Indicator<Num> {
      */
     public int getSize() {
         return barSeries.getBarCount() - 1;
-    }
-
-    public void calculate(Position position) {
-        calculate(position, barSeries.getEndIndex());
     }
 
     /**
@@ -212,15 +208,16 @@ public class Returns implements Indicator<Num> {
      * @param tradingRecord the trading record
      */
     private void calculate(TradingRecord tradingRecord) {
+        int endIndex = tradingRecord.getEndIndex(getBarSeries());
         // For each position...
-        tradingRecord.getPositions().forEach(this::calculate);
+        tradingRecord.getPositions().forEach(p -> calculate(p, endIndex));
     }
 
     /**
-     * Fills with zeroes until the end of the series.
+     * Fills with zeroes until the endIndex.
      */
-    private void fillToTheEnd() {
-        if (barSeries.getEndIndex() >= values.size()) {
+    private void fillToTheEnd(int endIndex) {
+        if (endIndex >= values.size()) {
             values.addAll(Collections.nCopies(barSeries.getEndIndex() - values.size() + 1, barSeries.numOf(0)));
         }
     }