+++ ta4j-core/src/main/java/org/ta4j/core/criteria/helpers/AverageCriterion.java
@@ -35,7 +35,8 @@ import org.ta4j.core.num.Num;
  * Average criterion.
  * 
  * <p>
- * Calculates the average of a Criterion.
+ * Calculates the average of a Criterion by dividing it by the number of
+ * positions.
  */
 public class AverageCriterion extends AbstractAnalysisCriterion {